pyproject.toml
[project]
name = "paraplan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.5",
    "orjson>=3.10.7",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "telegrinder>=0.2.2",
    "jinja2>=3.1.4",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "sqlalchemy>=2.0.37",
]

[tool.uv]
dev-dependencies = ["pyright>=1.1.379", "ruff>=0.6.4"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
lint.select = ["I", "E", "F", "B", "Q"]
lint.ignore = ["B008"]
line-length = 99
